refactor: Queue<Direction>을 생성하는 기능을 Directions로 이동
- Queue<Direction>을 생성하는 기능은 Directions에 있는 것이 더 자연스러움
- direction 패키지 생성 후 Direction, Directions 추가
- direction 테스트 패키지도 생성
- Position에서 Direction만 의존하게 변경
- 위의 과정에 의한 수정사항들 변경
