+++ src/main/java/domain/piece/strategy/ChariotMovingCondition.java
@@ -1,13 +1,14 @@
 package domain.piece.strategy;
 
 import domain.board.BoardState;
+import domain.direction.Directions;
 import domain.position.Position;
 
 public class ChariotMovingCondition implements MovingCondition {
 
     @Override
     public boolean canMove(BoardState boardState, Position startPosition, Position endPosition) {
-        Directions directions = Directions.between(startPosition, endPosition);
+        Directions directions = Directions.of(startPosition, endPosition);
         if (!directions.checkAllDirectionIsStraight()) {
             return false;
         }
