+++ src/main/java/domain/piece/strategy/ElephantMovingCondition.java
@@ -1,16 +1,16 @@
 package domain.piece.strategy;
 
 import domain.board.BoardState;
-import domain.board.Direction;
+import domain.direction.Direction;
+import domain.direction.Directions;
 import domain.position.Position;
-import domain.position.PositionCalculator;
 
 public class ElephantMovingCondition implements MovingCondition {
     private static final int MAX_DIRECTION = 3;
 
     @Override
     public boolean canMove(BoardState boardState, Position startPosition, Position endPosition) {
-        Directions directions = Directions.between(startPosition, endPosition);
+        Directions directions = Directions.of(startPosition, endPosition);
 
         if (directions.size() != MAX_DIRECTION) {
             return false;
@@ -26,11 +26,11 @@ public class ElephantMovingCondition implements MovingCondition {
         return firstDirection.isStraight()
                 && isNotBlocked(boardState, startPosition, firstDirection)
                 && secondDirection.isSameAtLeastOne(firstDirection)
-                && isNotBlocked(boardState, PositionCalculator.add(startPosition, firstDirection), secondDirection)
+                && isNotBlocked(boardState, startPosition.append(firstDirection), secondDirection)
                 && secondDirection.isSameDirection(thirdDirection);
     }
 
     private boolean isNotBlocked(BoardState boardState, Position position, Direction direction) {
-        return !boardState.isBlocked(PositionCalculator.add(position, direction));
+        return !boardState.isBlocked(position.append(direction));
     }
 }
