+++ src/main/java/domain/piece/strategy/HorseMovingCondition.java
@@ -1,9 +1,9 @@
 package domain.piece.strategy;
 
 import domain.board.BoardState;
-import domain.board.Direction;
+import domain.direction.Direction;
+import domain.direction.Directions;
 import domain.position.Position;
-import domain.position.PositionCalculator;
 
 public class HorseMovingCondition implements MovingCondition {
 
@@ -11,7 +11,7 @@ public class HorseMovingCondition implements MovingCondition {
 
     @Override
     public boolean canMove(BoardState boardState, Position startPosition, Position endPosition) {
-        Directions directions = Directions.between(startPosition, endPosition);
+        Directions directions = Directions.of(startPosition, endPosition);
 
         if (directions.size() != MAX_DIRECTION) {
             return false;
@@ -29,6 +29,6 @@ public class HorseMovingCondition implements MovingCondition {
     }
 
     private boolean isNotBlocked(BoardState boardState, Position position, Direction direction) {
-        return !boardState.isBlocked(PositionCalculator.add(position, direction));
+        return !boardState.isBlocked(position.append(direction));
     }
 }
