+++ src/main/java/domain/piece/strategy/PawnMovingCondition.java
@@ -1,7 +1,8 @@
 package domain.piece.strategy;
 
 import domain.board.BoardState;
-import domain.board.Direction;
+import domain.direction.Direction;
+import domain.direction.Directions;
 import domain.position.Position;
 
 public class PawnMovingCondition implements MovingCondition {
@@ -10,7 +11,7 @@ public class PawnMovingCondition implements MovingCondition {
 
     @Override
     public boolean canMove(BoardState boardState, Position startPosition, Position endPosition) {
-        Directions directions = Directions.between(startPosition, endPosition);
+        Directions directions = Directions.of(startPosition, endPosition);
         if (directions.size() != MAX_DIRECTION) {
             return false;
         }
